# api/openapi/helm.openapi.yaml
openapi: "3.1.0"
info:
  title: HELM Kernel API
  version: 0.5.0
  description: |
    Deterministic execution kernel for AI tool calls.
    Drop-in OpenAI proxy + cryptographic receipts + offline-verifiable evidence packs.
  license:
    name: Apache-2.0
    url: https://github.com/Mindburn-Labs/helm-ai-kernel/blob/main/LICENSE
  contact:
    name: Mindburn Labs
    url: https://mindburn.org
servers:
  - url: http://localhost:8080
    description: Local development (Docker Compose)
  - url: http://localhost:9090
    description: Standalone proxy mode
...
paths:
...
  /v1/chat/completions:
    post:
      operationId: chatCompletions
      tags: [proxy]
      summary: OpenAI-compatible chat completions
      description: |
        Drop-in replacement for OpenAI /v1/chat/completions.
        Tool calls are intercepted, schema-validated, and receipted.
        Response includes X-Helm-Receipt-ID, X-Helm-Output-Hash, X-Helm-Lamport-Clock headers.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ChatCompletionRequest'
            example:
              model: gpt-4
              messages:
                - role: user
                  content: "List files in /tmp"
              tools:
                - type: function
                  function:
                    name: file_read
                    parameters:
                      type: object
                      properties:
                        path:
                          type: string
      responses:
        '200':
          description: Chat completion with optional tool calls
          headers:
            X-Helm-Receipt-ID:
              schema: { type: string }
              description: Unique receipt ID for this execution
            X-Helm-Output-Hash:
              schema: { type: string }
              description: SHA-256 of the response body
            X-Helm-Lamport-Clock:
              schema: { type: integer }
              description: Causal ordering clock
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ChatCompletionResponse'
        '400':
          $ref: '#/components/responses/HelmError'
        '403':
          $ref: '#/components/responses/HelmError'
        '500':
          $ref: '#/components/responses/HelmError'

  # ── Kernel Approval ─────────────────────────────────
...
  /api/v1/evaluate:
    post:
      operationId: evaluateDecision
      tags: [console]
      summary: Evaluate a governed intent and emit a receipt
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/DecisionRequest'
      responses:
        '200':
          description: Guardian decision record
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DecisionRecord'
        '400':
          $ref: '#/components/responses/HelmError'
        '500':
          $ref: '#/components/responses/HelmError'
  /api/v1/receipts:
    get:
      operationId: listReceipts
      tags: [console]
      summary: List durable receipts
      security:
        - AdminBearerAuth: []
      parameters:
        - name: limit
          in: query
          schema: { type: integer, default: 100, maximum: 1000 }
        - name: since
          in: query
          schema: { type: string }
        - name: agent
          in: query
          schema: { type: string }
      responses:
        '200':
          description: Receipt list
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ReceiptListResponse'
        '400':
          $ref: '#/components/responses/HelmError'
        '401':
          $ref: '#/components/responses/HelmError'
        '503':
          $ref: '#/components/responses/HelmError'
...
components:
  securitySchemes:
    AdminBearerAuth:
      type: http
      scheme: bearer
      bearerFormat: HELM_ADMIN_API_KEY
      description: |
        Self-hosted OSS admin bearer key. Set HELM_ADMIN_API_KEY on the server
        and send Authorization: Bearer <key> for protected runtime APIs.

  # ── Schemas ─────────────────────────────────────────
  schemas:
    HelmError:
      type: object
      required: [error]
      properties:
        error:
          type: object
          required: [message, type, code, reason_code]
          properties:
            message:
              type: string
              description: Human-readable error message
            type:
              type: string
              enum: [invalid_request, authentication_error, permission_denied, not_found, internal_error]
            code:
              type: string
              description: Machine-readable error code
            reason_code:
              type: string
              description: HELM-specific reason code
              enum:
                - DENY_TOOL_NOT_FOUND
                - DENY_SCHEMA_MISMATCH
                - DENY_OUTPUT_DRIFT
                - DENY_BUDGET_EXCEEDED
                - DENY_APPROVAL_REQUIRED
                - DENY_APPROVAL_TIMEOUT
                - DENY_SANDBOX_TRAP
                - DENY_GAS_EXHAUSTION
                - DENY_TIME_LIMIT
                - DENY_MEMORY_LIMIT
                - DENY_POLICY_VIOLATION
                - DENY_TRUST_KEY_REVOKED
                - DENY_IDEMPOTENCY_DUPLICATE
                - ERROR_INTERNAL
            details:
              type: object
              description: Additional error context
              additionalProperties: true
    ChatCompletionRequest:
      type: object
      required: [model, messages]
      properties:
        model:
          type: string
        messages:
          type: array
          items:
            type: object
            required: [role, content]
            properties:
              role:
                type: string
                enum: [system, user, assistant, tool]
              content:
                type: string
              tool_call_id:
                type: string
        tools:
          type: array
          items:
            type: object
            properties:
              type: { type: string, enum: [function] }
              function:
                type: object
                properties:
                  name: { type: string }
                  description: { type: string }
                  parameters: { type: object }
        temperature:
          type: number
        max_tokens:
          type: integer
        stream:
          type: boolean
          default: false
    ChatCompletionResponse:
      type: object
      properties:
        id: { type: string }
        object: { type: string, default: "chat.completion" }
        created: { type: integer }
        model: { type: string }
        choices:
          type: array
          items:
            type: object
            properties:
              index: { type: integer }
              message:
                type: object
                properties:
                  role: { type: string }
                  content:
                    type: ['string', 'null']
                  tool_calls:
                    type: array
                    items:
                      type: object
                      properties:
                        id: { type: string }
                        type: { type: string }
                        function:
                          type: object
                          properties:
                            name: { type: string }
                            arguments: { type: string }
              finish_reason: { type: string }
        usage:
          type: object
          properties:
            prompt_tokens: { type: integer }
            completion_tokens: { type: integer }
            total_tokens: { type: integer }
...
    Receipt:
      type: object
      properties:
        receipt_id: { type: string }
        decision_id: { type: string }
        effect_id: { type: string }
        status: { type: string }
        reason_code: { type: string }
        output_hash: { type: string }
        blob_hash: { type: string }
        prev_hash: { type: string }
        lamport_clock: { type: integer }
        signature: { type: string }
        timestamp: { type: string, format: date-time }
        principal: { type: string }
        executor_id: { type: string }
        args_hash: { type: string }
        metadata:
          type: object
          additionalProperties: true
    ReceiptListResponse:
      type: object
      properties:
        receipts:
          type: array
          items:
            $ref: '#/components/schemas/Receipt'
        count: { type: integer }
        next_cursor: { type: string }
        has_more: { type: boolean }
    DecisionRequest:
      type: object
      required: [action, resource]
      properties:
        principal: { type: string }
        action: { type: string }
        resource: { type: string }
        context:
          type: object
          additionalProperties: true
    DecisionRecord:
      type: object
      properties:
        id: { type: string }
        action: { type: string }
        resource: { type: string }
        verdict: { type: string }
        reason: { type: string }
        policy_version: { type: string }
        policy_decision_hash: { type: string }
        signature: { type: string }
...
  responses:
    HelmError:
      description: HELM error response
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/HelmError'
          example:
            error:
              message: "Tool not found: unknown_tool"
              type: "invalid_request"
              code: "tool_not_found"
              reason_code: "DENY_TOOL_NOT_FOUND"